speed update
First attempt to make the data load faster.  Loading everything in one get of the match instead of 2.  Thinking it will run slow for the first load of a champion but will cache a lot of the data.  Still have not decided on whether to store the pictures locally or not.